Refactored old queries structure for transactions
