--- src/pages/transactions/common/hooks/useBulkAction.tsx
@@ -1,42 +1,0 @@
-/**
- * Invoice Ninja (https://invoiceninja.com).
- *
- * @link https://github.com/invoiceninja/invoiceninja source repository
- *
- * @copyright Copyright (c) 2022. Invoice Ninja LLC (https://invoiceninja.com)
- *
- * @license https://www.elastic.co/licensing/elastic-license
- */
-
-import { endpoint } from '$app/common/helpers';
-import { request } from '$app/common/helpers/request';
-import { toast } from '$app/common/helpers/toast/toast';
-import { useQueryClient } from 'react-query';
-import { route } from '$app/common/helpers/route';
-import { useAtomValue } from 'jotai';
-import { invalidationQueryAtom } from '$app/common/atoms/data-table';
-
-export function useBulkAction() {
-  const queryClient = useQueryClient();
-  const invalidateQueryValue = useAtomValue(invalidationQueryAtom);
-
-  return (id: string, action: 'archive' | 'restore' | 'delete') => {
-    toast.processing();
-
-    request('POST', endpoint('/api/v1/bank_transactions/bulk'), {
-      action,
-      ids: [id],
-    }).then(() => {
-      toast.success(`${action}d_transaction`);
-
-      queryClient.invalidateQueries('/api/v1/bank_transactions');
-
-      queryClient.invalidateQueries(
-        route('/api/v1/bank_transactions/:id', { id })
-      );
-
-      invalidateQueryValue &&
-        queryClient.invalidateQueries([invalidateQueryValue]);
-    });
-  };
-}
