+++ src/pages/transactions/create/Create.tsx
@@ -24,13 +24,13 @@ import { useCurrentCompany } from '$app/common/hooks/useCurrentCompany';
 import { DecimalInputSeparators } from '$app/common/interfaces/decimal-number-input-separators';
 import { ApiTransactionType } from '$app/common/enums/transactions';
 import { ValidationBag } from '$app/common/interfaces/validation-bag';
-import { useBlankTransactionQuery } from '../common/queries';
 import { useResolveCurrencySeparator } from '../common/hooks/useResolveCurrencySeparator';
 import { TransactionForm } from '../components/TransactionForm';
 import { useHandleChange } from '../common/hooks/useHandleChange';
 import { GenericSingleResourceResponse } from '$app/common/interfaces/generic-api-response';
 import { route } from '$app/common/helpers/route';
 import { useQueryClient } from 'react-query';
+import { useBlankTransactionQuery } from '$app/common/queries/transactions';
 
 export default function Create() {
   const [t] = useTranslation();
